proje_yonetimi/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
google-auth>=2.29